requirements.txt
Flask
Pillow
pandas
openpyxl